Move growthbook setup to growthbook.ts and minor requested changes
