+++ projects/bp-gallery/src/components/common/IfFlagEnabled.tsx
@@ -1,12 +1,9 @@
 import { IfFeatureEnabled } from '@growthbook/growthbook-react';
 import { PropsWithChildren } from 'react';
-import { FeatureId } from '../../helpers/growthbook';
-
-const growthbookApiHost = import.meta.env.VITE_REACT_APP_GROWTHBOOK_APIHOST;
-const growthbookClientKey = import.meta.env.VITE_REACT_APP_GROWTHBOOK_CLIENTKEY;
+import { FeatureId, growthbook } from '../../helpers/growthbook';
 
 export const IfFlagEnabled = ({ children, feature }: PropsWithChildren<{ feature: FeatureId }>) => {
-  if (growthbookApiHost && growthbookClientKey) {
+  if (growthbook) {
     return <IfFeatureEnabled feature={feature}>{children}</IfFeatureEnabled>;
   }
   return <>{children}</>;
